Fixed harmless type errors

+++ packages/xpce/src/ker/alloc.c
@@ -124,7 +124,7 @@ about  the same  relative numbers.    For  this reason  PCE addopts  a
 perfect fit strategy for memory allocation.
 - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 
-#define offset(structure, field) ((int) &(((structure *)NULL)->field))
+#define offset(structure, field) ((intptr_t) &(((structure *)NULL)->field))
 
 static inline Zone
 allocate(size_t size)
